% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{.Petfinder.class}
\alias{.Petfinder.class}
\title{Internal R6 class}
\description{
Internal R6 class

Internal R6 class
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{.Petfinder.class$new()}}
\item \href{#method-animal_types}{\code{.Petfinder.class$animal_types()}}
\item \href{#method-breeds}{\code{.Petfinder.class$breeds()}}
\item \href{#method-animals}{\code{.Petfinder.class$animals()}}
\item \href{#method-organizations}{\code{.Petfinder.class$organizations()}}
\item \href{#method-clone}{\code{.Petfinder.class$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{.Petfinder.class$new(key, secret)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-animal_types"></a>}}
\if{latex}{\out{\hypertarget{method-animal_types}{}}}
\subsection{Method \code{animal_types()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{.Petfinder.class$animal_types(types = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-breeds"></a>}}
\if{latex}{\out{\hypertarget{method-breeds}{}}}
\subsection{Method \code{breeds()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{.Petfinder.class$breeds(animal_types = NULL, return_df = FALSE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-animals"></a>}}
\if{latex}{\out{\hypertarget{method-animals}{}}}
\subsection{Method \code{animals()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{.Petfinder.class$animals(
  animal_id = NULL,
  animal_type = NULL,
  breed = NULL,
  size = NULL,
  gender = NULL,
  age = NULL,
  color = NULL,
  coat = NULL,
  status = NULL,
  name = NULL,
  organization = NULL,
  location = NULL,
  distance = NULL,
  sort = NULL,
  good_with_cats = NULL,
  good_with_children = NULL,
  good_with_dogs = NULL,
  before_date = NULL,
  after_date = NULL,
  pages = 1,
  results_per_page = 20,
  return_df = FALSE
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-organizations"></a>}}
\if{latex}{\out{\hypertarget{method-organizations}{}}}
\subsection{Method \code{organizations()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{.Petfinder.class$organizations(
  organization_id = NULL,
  name = NULL,
  location = NULL,
  distance = NULL,
  state = NULL,
  country = NULL,
  query = NULL,
  sort = NULL,
  results_per_page = 20,
  pages = 1,
  return_df = FALSE
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{.Petfinder.class$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
