% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetAnimals.R
\name{GetAnimals}
\alias{GetAnimals}
\title{Access All Available Petfinder Data}
\usage{
GetAnimals(
  breed = NULL,
  size = NULL,
  gender = NULL,
  age = NULL,
  color = NULL,
  coat = NULL,
  status = NULL,
  organization = NULL,
  good_with_children = NULL,
  good_with_dogs = NULL,
  good_with_cats = NULL,
  location = NULL,
  distance = NULL,
  before = NULL,
  after = NULL,
  limit = 20
)
}
\arguments{
\item{breed}{Accepts multiple values, e.g. breed=pug,samoyed.}

\item{size}{small, medium, large, xlarge Accepts multiple values, e.g. size=large,xlarge.}

\item{gender}{male, female, unknown Accepts multiple values, e.g. gender=male,female.}

\item{age}{baby, young, adult, senior Accepts multiple values, e.g. age=baby,senior.}

\item{color}{Return results matching animal color}

\item{coat}{Return results matching animal coat}

\item{status}{Return results matching adoption status}

\item{organization}{Return results associated with specific organization(s)}

\item{good_with_children}{Return results that are good with children. Can be true, false, 1, or 0}

\item{good_with_dogs}{Return results that are good with dogs. Can be true, false, 1, or 0}

\item{good_with_cats}{Return results that are good with cats. Can be true, false, 1, or 0}

\item{location}{Return results by location. string city, state; latitude,longitude; or postal code.}

\item{distance}{Return results within distance of location (in miles). Requires location to be set (default: 100, max: 500)}

\item{before}{Return results published before this date/time. Must be a valid ISO8601 date-time string (e.g. 2019-10-07T19:13:01+00:00)}

\item{after}{Return results published after this date/time. Must be a valid ISO8601 date-time string (e.g. 2019-10-07T19:13:01+00:00)}

\item{limit}{Sets the number of observations to return. Default is 20.}
}
\description{
This function allows the users to access pet adoption data from Petfinder 2.0 API
through a variety of different parameters. Parameters should be left null if it is not of interest
so as to include all fields that have that parameter.
}
\details{
Returns a dataframe with all available features for 20 observations.
Parameters will filter based on selected values and number of observations can be adjusted with
the limit parameter.
}
\examples{
GetAnimals()

}
