% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/petfindeR-package.R
\docType{package}
\name{petfindeR-package}
\alias{petfindeR}
\alias{petfindeR-package}
\title{petfindeR: Petfinder API Client}
\description{
The petfinder API Client allows R users to access adoption data for animals across shelters in the united states.
}
\keyword{internal}
