% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method_docs.R
\name{shelter.get}
\alias{shelter.get}
\title{Returns shelter records of input shelterIds.}
\usage{
shelter.get(shelterId)
}
\arguments{
\item{shelterId}{ID of the shelter record to return. Accepts a character for
a single record or a list or vector of shelterIds.}
}
\value{
data.frame of input shelterId(s).
}
\description{
Returns shelter records of input shelterIds.
}
\examples{
pf <- Petfinder(key)
pf$shelter.get(shelterId)
shelterIds <- c(shelterId1, shelterId2, shelterId3)
pf$shelter.get(shelterIds)
}
