% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{breed.list}
\alias{breed.list}
\title{Method for calling the 'breed.list' method of the Petfinder API. Returns the
available breeds for the selected animal.}
\usage{
breed.list(animal, outputformat = "json")
}
\arguments{
\item{animal}{Return breeds of animal. Must be one of 'barnyard', 'bird',
'cat', 'dog', 'horse', 'reptile', or 'smallfurry'}

\item{outputformat}{Output type of results. Must be one of 'json' (default)
or 'xml'}
}
\value{
The breeds of the animal. If the parameter :code:`outputformat` is
  'json', the result is formatted as a JSON object. Otherwise, the return
  object is a text representation of an XML object.
}
\description{
Method for calling the 'breed.list' method of the Petfinder API. Returns the
available breeds for the selected animal.
}
\examples{
pf <- Petfinder(key)
pf$breed.list('cat')
pf$breed.list('dog', 'xml')
}
