% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method_docs.R
\name{breed.list}
\alias{breed.list}
\title{Method for calling the 'breed.list' method of the Petfinder API. Returns the
available breeds for the selected animal.}
\usage{
breed.list(animal)
}
\arguments{
\item{animal}{Return breeds of animal. Must be one of 'barnyard', 'bird',
'cat', 'dog', 'horse', 'reptile', or 'smallfurry'.}
}
\value{
data.frame containing available breeds of the animal from the Petfinder database.
}
\description{
Method for calling the 'breed.list' method of the Petfinder API. Returns the
available breeds for the selected animal.
}
\examples{
\dontrun{
pf <- Petfinder(key) # Initialize Petfinder class
pf$breed.list('cat')
pf$breed.list('dog')
}
}
