% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\docType{data}
\name{.Petfinder.class}
\alias{.Petfinder.class}
\title{Internal R6 class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
.Petfinder.class
}
\description{
Internal R6 class
}
\keyword{datasets}
