% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method_docs.R
\name{pet.get}
\alias{pet.get}
\title{Returns the available pet record data for the input pet Ids.}
\usage{
pet.get(petId)
}
\arguments{
\item{petId}{ID of the pet record to return. Accepts a character for a single
record or a list or vector of petIds.}
}
\value{
data.frame of input petId(s).
}
\description{
Returns the available pet record data for the input pet Ids.
}
\examples{
pf <- Petfinder(key)
pf$pet.get(petId)
petIds <- c(petId1, petId2, petId3)
pf$pet.get(petIds)
}
